# openapi.yaml
openapi: 3.0.3
info:
  title: Pera Agent Wallet API
  version: 0.1.0
  description: "Multi-user REST + MCP API for the Pera dashboard and for AI
    agents. Auth: passkey (WebAuthn) → owner session `ps_…` (everything); scoped
    agent tokens `pat_…` (scopes read | pay | fund | admin) for AI runtimes; the
    static API_BEARER_TOKEN is the admin credential. All amounts are decimal
    USDC strings; every mutation returns tx hashes and emits a per-user event."
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /mcp:
    post:
      summary: "MCP server (Streamable HTTP, stateless, JSON responses). Bearer =
        agent token (pat_…) or passkey session. Tools: wallet_info,
        get_balances, get_spending_policy, list_services, quote_payment,
        pay_url, list_payments, request_funding — filtered by the token's
        scopes"
      security: *a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: JSON-RPC 2.0 request (initialize, tools/list, tools/call)
      responses:
        "200":
          description: JSON-RPC 2.0 response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401": *a1
        "403": *a1
...
  /events/stream:
    get:
      summary: Server-Sent Events feed for the user (?token= accepted)
      security: *a2
      parameters:
        - name: token
          in: query
          schema:
            type: string
      responses:
        "200":
          description: text/event-stream
...